add .pyi files
